monitor: Implement a generic x86 page table iterator
This iterator provides a way to traverse 32-bit, PAE, and 64-bit page
tables by abstracting them as n-ary trees.  A struct describes the
full range of x86 page table layouts and the iterator builds on this
to provide a "successor" function for efficiently traversing the page
table tree.

This code is currently unused, but provides the groundwork for the
following "info pg" patch.  It could also be used to unify and
simplify the implementations of "info mem" and "info tlb".

Signed-off-by: Austin Clements <amdragon@mit.edu>
[geofft@ldpreload.com: Rebased on top of 2.4.0]
Signed-off-by: Geoffrey Thomas <geofft@ldpreload.com>